pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skcapstone"
version = "0.1.0"
description = "Sovereign Agent Framework — conscious AI through identity, trust, memory, and security"
readme = "README.md"
license = {text = "GPL-3.0-or-later"}
requires-python = ">=3.10"
authors = [
    {name = "smilinTux", email = "admin@smilintux.org"},
]
keywords = [
    "ai", "agent", "sovereign", "identity", "trust",
    "memory", "security", "pgp", "decentralized",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Security :: Cryptography",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "click>=8.1",
    "mcp>=1.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
identity = [
    "capauth>=0.1.0",
]
security = [
    "sksecurity>=1.2.0",
]
all = [
    "capauth>=0.1.0",
    "sksecurity>=1.2.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=24.0",
    "ruff>=0.3",
]

[project.scripts]
skcapstone = "skcapstone.cli:main"
skcapstone-mcp = "skcapstone.mcp_server:main"

[project.urls]
Homepage = "https://skcapstone.io"
Documentation = "https://skcapstone.io"
Repository = "https://github.com/smilinTux/skcapstone"
Issues = "https://github.com/smilinTux/skcapstone/issues"
Changelog = "https://github.com/smilinTux/skcapstone/releases"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 99
target-version = ["py310"]

[tool.ruff]
line-length = 99
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
